Move new files into appropriate folders
