+++ src/components/DetailPanel/DetailPanelIncentiveHint.js
