+++ src/components/DetailPanel/DetailPanelMain.js
