+++ src/components/icons/ui-elements/AddPhotoIcon.js
